Move serial error enum to rhsplib
