+++ packages/rhsplib/lib/serial-errors.ts
@@ -1,0 +1,7 @@
+export enum SerialError {
+    GENERAL_ERROR = -101,
+    UNABLE_TO_OPEN = -102,
+    INVALID_ARGS = -103,
+    CONFIGURATION_ERROR = -104,
+    IO_ERROR = -105,
+}
